refactor(user-profile): split profile modules (WP3.6)
