+++ docs/REFACTOR_PLAN.md
@@ -1,8 +1,9 @@
 # Deep Refactor Plan — v3 (2026-07-04, full-scan grounded)
 
-**Status: Track A, Phase -1, and Phase 1 (WP1.1–WP1.8) completed; Track B
-mostly shipped (WPB.3 and WPB.6 landed 2026-07-07; WPB.4 remains gated); Plan-v3
-structural-refactor execution approved by the owner on 2026-07-05. Phase 2 is next.**
+**Status: Track A and Phases -1 through 2 completed; Phase 3 is complete through
+the optional WP3.6 split in the current working tree on `main` @ `8c6acb6`. Track B is
+mostly shipped; WPB.4 remains gated on product-risk review. In the isolated Phase-4
+worktree, WP4.-1 and WP4.0a are committed and WP4.0 is in progress.**
 
 This supersedes v2. It incorporates:
 
@@ -307,6 +308,26 @@ Playwright inventory **504 tests / 30 specs**.
 - Gate: weekly-summary pytest + goldens diff reviewed as intentional, plus
   weekly-summary visual/E2E specs.
 
+**Risk-mitigation gate (reviewed 2026-07-17):**
+
+- The production schema makes `routine` `TEXT NOT NULL`; the reachable case is an empty
+  string, while `None` remains relevant to mocked/legacy rows. State explicitly that all
+  falsy routine values coalesce into one synthetic `Unassigned` session.
+- Freeze scope to session-derived metrics only: weekly raw/effective totals, reps, volume,
+  status, contribution weights, rounding, response fields, and pattern coverage must not
+  change. Decide separately whether `global_sessions` includes the synthetic bucket; do
+  not let that denominator change happen implicitly.
+- Add focused cases for empty/`None`, above- and below-1.0 frequency thresholds, multiple
+  anonymous rows accumulating into one bucket, and mixed named/anonymous routines across
+  the full counting-mode x contribution-mode matrix.
+- Regenerate the WP2.3 golden only after reviewing the exact delta. For the existing Calves
+  sentinel, the intended change is frequency `0 -> 1`; effective-mode `sets_per_session`
+  `0.85 -> 5.1`; raw-mode `sets_per_session` `1 -> 6`; and effective-derived average/max
+  `0/0 -> 5.1/5.1`. Weekly totals and classifications must remain identical.
+- Add route/E2E assertions for the displayed frequency and average/max-per-session values,
+  run summary-page functional and visual gates, and include migration notes describing the
+  intentional semantic change and the conservative one-bucket assumption.
+
 ### WPB.5 (OD5) Experience-aware increment below 20 kg
 
 - Protected calc zone. In `_calculate_weight_increment`, experienced lifters get
@@ -595,6 +616,12 @@ file-to-package conversion.
   separate internal-caller proof authorizes it. The unused loop variable may be cleaned.
 - Gate: unmodified plan-generator tests plus starter-plan E2E.
 
+**Completed 2026-07-16.** Extracted scoring, priority-allocation, persistence, and
+result-assembly helpers without changing the public callable signature, score ordering,
+row-order mutation, or the inner-continue/outer-reraise exception tiers. Added explicit
+contract tests for those seams. Local gate: **1,723 pytest passed** and the complete
+API-integration + workout-plan Chromium pair **92 passed**.
+
 ### WP2.3 Weekly-summary decomposition with durable goldens
 
 - First commit deterministic seeded golden fixtures for both public calculations,
@@ -739,6 +766,15 @@ replace-exercise-errors E2E after every boundary-affecting move.
 Only after the core JS track: characterize and split the 1,483-line file by demographics,
 reference lifts, coverage/bodymap, calibration review, and settings toggles. Consolidating
 the two optimistic-toggle paths is a later behavior-aware cleanup, not part of move-only PRs.
+
+**Completed 2026-07-17 in the current working tree.** The original entry module is now a
+small coordinator over focused data, forms/autosave, insights, bodymap, settings, and
+calibration-review modules. Initialization order, DOM hooks, API endpoints, payloads,
+toasts, rollback behavior, and the two distinct optimistic-toggle implementations are
+unchanged. Added pure estimator-seam characterization tests. Gate: Vitest **105 passed**,
+focused Python **75 passed**, full pytest **1,723 passed**, and profile +
+learned-calibration + fatigue-context Chromium **38 passed** against the isolated E2E
+database.
 
 ---
